+++ Backend/TekusServer/Application/Commands/ServicesProvider/Delete/DeleteServiceHandler.cs
@@ -1,10 +1,12 @@
-﻿using Application.Abstractions;
-using Application.Abstractions.Persistence;
+﻿using Application.Abstractions.Persistence;
 using Domain.Common;
+using Domain.Exceptions;
+using MediatR;
 
 namespace Application.Commands.ServicesProvider.Delete
 {
-    internal sealed class DeleteServiceHandler : ICommandHandler<DeleteServiceCommand>
+
+    internal sealed class DeleteServiceHandler : IRequestHandler<DeleteServiceCommand>
     {
         private readonly IServiceRepository _serviceRepository;
         private readonly IUnitOfWork _unitOfWork;
@@ -15,10 +17,12 @@ namespace Application.Commands.ServicesProvider.Delete
             _unitOfWork = unitOfWork;
         }
 
-        public async Task HandleAsync(DeleteServiceCommand command, CancellationToken cancellationToken = default)
+        public async Task Handle(DeleteServiceCommand command, CancellationToken cancellationToken)
         {
             var service = await _serviceRepository.GetById(command.Id, cancellationToken);
-            //if (service is null) throw new NotFoundException(...);
+
+            if (service == null)
+                throw new DomainException($"Service with ID {command.Id} was not found.");
 
             _serviceRepository.Delete(service);
             await _unitOfWork.SaveChanges(cancellationToken);
